+++ src/components/main/main.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Main from './main.jsx';
 import renderer from 'react-test-renderer';
-import {rentNames} from '../../testdata.js';
+import {rentNames} from '../../utils/testdata.js';
 
 describe(`Main`, () => {
   it(`Main component should render correctly`, () => {
